P5-7+P5-8: Decompose flow-view + note-editor god components
flow-view.tsx: 1019→862 LOC — extracted pure functions (computeGroupAssignments,
flowNodeToRF, flowEdgeToRF, defaultData, type registries, ADD_NODE_MENU,
CONNECTION_LINE_STYLE) into flow-utils.ts (146 LOC). Also fixed hardcoded
#6366f1 → var(--accent) and text-white → text-[var(--background)].

note-editor.tsx: 1283→932 LOC — extracted 3 sub-components:
- BacklinksPanel.tsx (225 LOC) — BacklinksPanel + NoteTagBar
- MDPreviewPanel.tsx (97 LOC) — markdown preview panel
- note-editor-utils.ts (31 LOC) — countWords + stripMarkdown helpers

type-check:all, lint, 430 tests pass.
