+++ src/components/notes/note-editor.tsx
@@ -7,14 +7,13 @@ import remarkBreaks from "remark-breaks";
 import remarkMath from "remark-math";
 import rehypeKatex from "rehype-katex";
 import "katex/dist/katex.min.css";
-import { Pin, PinOff, Calendar, Eye, Pencil, Wand2, Loader2, CheckCircle2, Tag, Plus, X, Link2, Kanban, ChevronDown, FileText, FileDown, ChevronLeft } from "lucide-react";
+import { Pin, PinOff, Calendar, Eye, Pencil, Wand2, Loader2, CheckCircle2, FileDown, ChevronLeft } from "lucide-react";
 import { WikilinkPicker } from "./WikilinkPicker";
 import { getActiveWikilink } from "@/lib/wikilink-parser";
 import { useCairnStore } from "@/store";
 import { useShallow } from "zustand/react/shallow";
 import { cn, formatRelative, urlTransform } from "@/lib/utils";
 import { Tooltip } from "@/components/ui/tooltip";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import type { Note } from "@/types";
 import { MermaidDiagram } from "./MermaidDiagram";
@@ -25,102 +24,9 @@ import { MathBlock } from "./MathBlock";
 import { AITextToolbar, buildAIActionPrompt, applyFormat, type AITextAction, type FormatAction } from "./ai-text-toolbar";
 import { MarkdownEditor, type MarkdownEditorHandle } from "./markdown-editor";
 import { remarkCallout, remarkObsidianEmbeds, remarkWikilinks, remarkPromoteDisplayMath, makeLatexPlugins, InlineCode } from "@/lib/markdown/pipeline";
-
-// ── MD Preview Panel ──────────────────────────────────────────────────────────
-// Docked to the bottom of the editor. Renders selected raw markdown through
-// the full pipeline and displays it as a collapsible bottom panel.
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const PREVIEW_REMARK_PLUGINS: any[] = [remarkGfm, remarkBreaks, remarkMath, remarkPromoteDisplayMath, remarkCallout, remarkObsidianEmbeds];
-
-interface MDPreviewPanelProps {
-  text: string;
-  onDismiss: () => void;
-}
-
-function MDPreviewPanel({ text, onDismiss }: MDPreviewPanelProps) {
-  // Each panel instance gets its own latex plugin pair so the mutable capture
-  // array is isolated — safe because the panel remounts when text changes.
-  const { rehypeCaptureLatex: previewCapture, rehypeMergedPass: previewMerge } = useMemo(() => makeLatexPlugins(), []);
-
-  // Dismiss on Escape
-  useEffect(() => {
-    function onKey(e: KeyboardEvent) { if (e.key === "Escape") onDismiss(); }
-    document.addEventListener("keydown", onKey);
-    return () => document.removeEventListener("keydown", onKey);
-  }, [onDismiss]);
-
-  return (
-    <div
-      data-md-preview-portal
-      className="flex-shrink-0 border-t border-[var(--border)] animate-fade-in"
-      style={{ background: "var(--surface)" }}
-      onMouseDown={(e) => e.stopPropagation()}
-    >
-      {/* Header bar */}
-      <div className="flex items-center justify-between px-4 py-1.5 border-b border-[var(--border)]" style={{ background: "var(--surface-2)" }}>
-        <span className="text-[0.714rem] font-medium text-[var(--text-tertiary)] uppercase tracking-wide">Preview</span>
-        <button
-          onClick={onDismiss}
-          className="p-0.5 rounded text-[var(--text-tertiary)] hover:text-[var(--text-primary)] hover:bg-[var(--surface)] transition-colors"
-        >
-          <X size={11} />
-        </button>
-      </div>
-      {/* Rendered markdown — capped at ~30vh so it never swamps the editor */}
-      <div className="prose-cairn px-6 py-4 overflow-y-auto" style={{ maxHeight: "30vh" }}>
-        <ReactMarkdown
-          remarkPlugins={PREVIEW_REMARK_PLUGINS}
-          rehypePlugins={[previewCapture, rehypeKatex, previewMerge]}
-          urlTransform={urlTransform}
-          components={({
-            mark({ children }: React.HTMLAttributes<HTMLElement> & ExtraProps) {
-              return (
-                <mark className="rounded px-0.5" style={{ background: "color-mix(in srgb, var(--accent) 22%, transparent)", color: "var(--text-primary)" }}>
-                  {children}
-                </mark>
-              );
-            },
-            callout({ children, ...props }: React.HTMLAttributes<HTMLElement> & ExtraProps) {
-              const p = props as Record<string, string>;
-              return (
-                <Callout
-                  type={p["data-callout-type"] ?? "note"}
-                  title={p["data-title"] || undefined}
-                  collapsible={p["data-collapsible"] === "true"}
-                  defaultOpen={p["data-default-open"] !== "false"}
-                >
-                  {children}
-                </Callout>
-              );
-            },
-            mathblock({ children, ...props }: React.HTMLAttributes<HTMLElement> & ExtraProps) {
-              const p = props as Record<string, string>;
-              return <MathBlock renderedChildren={children} latex={p["data-latex"] ?? ""} />;
-            },
-            pre({ children }: React.HTMLAttributes<HTMLPreElement> & ExtraProps) {
-              const child = Array.isArray(children) ? children[0] : children;
-              const code = child as React.ReactElement<{ className?: string; children?: React.ReactNode }>;
-              const className = code?.props?.className ?? "";
-              const lang = className.replace("language-", "") || undefined;
-              const content = String(code?.props?.children ?? "").replace(/\n$/, "");
-              return <CodeBlock code={content} language={lang} />;
-            },
-            code({ children, className }: React.HTMLAttributes<HTMLElement> & ExtraProps) {
-              return <InlineCode className={className}>{children}</InlineCode>;
-            },
-          } as import("react-markdown").Components)}
-        >
-          {text}
-        </ReactMarkdown>
-      </div>
-    </div>
-  );
-}
-
-
-
-
+import { BacklinksPanel, NoteTagBar } from "./BacklinksPanel";
+import { MDPreviewPanel } from "./MDPreviewPanel";
+import { countWords, stripMarkdown } from "./note-editor-utils";
 
 interface NoteEditorProps {
   note: Note;
@@ -1021,263 +927,6 @@ export function NoteEditor({ note, onBack }: NoteEditorProps) {
           onDismiss={() => setPreviewText(null)}
         />
       )}
-    </div>
-  );
-}
-
-// ── Backlinks Panel ───────────────────────────────────────────────────────────
-
-interface BacklinksPanelProps {
-  note: Note;
-  onOpenCard: () => void;
-}
-
-function BacklinksPanel({ note, onOpenCard }: BacklinksPanelProps) {
-  const { notes, cards, columns } = useCairnStore(useShallow((s) => ({
-    notes:   s.notes,
-    cards:   s.cards,
-    columns: s.columns,
-  })));
-  const [open, setOpen] = useState(false);
-
-  const linkedNotes = useMemo(
-    () => (note.linkedNoteIds ?? []).map((id) => notes.find((n) => n.id === id)).filter(Boolean) as import("@/types").Note[],
-    [note.linkedNoteIds, notes],
-  );
-  const linkedCards = useMemo(
-    () => (note.linkedCardIds ?? []).map((id) => cards.find((c) => c.id === id)).filter(Boolean) as import("@/types").TaskCard[],
-    [note.linkedCardIds, cards],
-  );
-
-  // Notes that wikilink to this note via [[Current Note Title]]
-  const wikilinkBacklinks = useMemo(() => {
-    const titleLower = note.title.toLowerCase();
-    const re = /\[\[([^\][\n]+?)\]\]/g;
-    return notes.filter((n) => {
-      if (n.id === note.id) return false;
-      // Already shown in linkedNotes
-      if ((note.linkedNoteIds ?? []).includes(n.id)) return false;
-      const content = n.content ?? "";
-      let m: RegExpExecArray | null;
-      re.lastIndex = 0;
-      while ((m = re.exec(content)) !== null) {
-        if (m[1].trim().toLowerCase() === titleLower) return true;
-      }
-      return false;
-    });
-  }, [note.id, note.title, note.linkedNoteIds, notes]);
-
-  const total = linkedNotes.length + linkedCards.length + wikilinkBacklinks.length;
-  if (total === 0) return null;
-
-  return (
-    <div className="flex-shrink-0 border-t border-[var(--border)]">
-      <button
-        onClick={() => setOpen((o) => !o)}
-        className="w-full flex items-center gap-2 px-6 py-2.5 text-xs text-[var(--text-tertiary)] hover:text-[var(--text-secondary)] hover:bg-[var(--surface-2)] transition-colors"
-      >
-        <Link2 size={11} />
-        <span className="flex-1 text-left">{total} backlink{total !== 1 ? "s" : ""}</span>
-        <ChevronDown size={11} className={cn("transition-transform", open && "rotate-180")} />
-      </button>
-      {open && (
-        <div className="px-6 pb-3 space-y-1">
-          {linkedNotes.map((n) => (
-            <button
-              key={n.id}
-              onClick={() => window.dispatchEvent(new CustomEvent("cairn:select-note", { detail: { noteId: n.id } }))}
-              className="w-full flex items-center gap-2 px-2 py-1.5 rounded-md hover:bg-[var(--surface-2)] text-xs text-[var(--text-secondary)] transition-colors text-left"
-            >
-              <FileText size={11} className="text-[var(--text-tertiary)] flex-shrink-0" />
-              <span className="truncate flex-1">{n.title}</span>
-              <span className="text-[0.714rem] text-[var(--text-tertiary)]">note</span>
-            </button>
-          ))}
-          {linkedCards.map((c) => {
-            const col = columns.find((col) => col.id === c.columnId);
-            return (
-              <button
-                key={c.id}
-                onClick={onOpenCard}
-                className="w-full flex items-center gap-2 px-2 py-1.5 rounded-md hover:bg-[var(--surface-2)] text-xs text-[var(--text-secondary)] transition-colors text-left"
-              >
-                <Kanban size={11} className="text-[var(--text-tertiary)] flex-shrink-0" />
-                <span className="truncate flex-1">{c.title}</span>
-                <span className="text-[0.714rem] text-[var(--text-tertiary)]">{col?.name ?? "card"}</span>
-              </button>
-            );
-          })}
-          {wikilinkBacklinks.length > 0 && (
-            <>
-              {(linkedNotes.length > 0 || linkedCards.length > 0) && (
-                <div className="h-px bg-[var(--border)] my-1" />
-              )}
-              {wikilinkBacklinks.map((n) => (
-                <button
-                  key={n.id}
-                  onClick={() => window.dispatchEvent(new CustomEvent("cairn:select-note", { detail: { noteId: n.id } }))}
-                  className="w-full flex items-center gap-2 px-2 py-1.5 rounded-md hover:bg-[var(--surface-2)] text-xs text-[var(--text-secondary)] transition-colors text-left"
-                >
-                  <Link2 size={11} className="text-[var(--accent)] flex-shrink-0" />
-                  <span className="truncate flex-1">{n.title}</span>
-                  <span className="text-[0.714rem] text-[var(--accent)] opacity-70">wikilink</span>
-                </button>
-              ))}
-            </>
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ── Note Tag Bar ──────────────────────────────────────────────────────────────
-
-interface NoteTagBarProps {
-  note: Note;
-  workspaceTags: import("@/types").Tag[];
-  onToggleTag: (tagId: string) => void;
-  onCreateTag: (name: string) => void;
-  getTagById: (id: string) => import("@/types").Tag | undefined;
-}
-
-function NoteTagBar({ note, workspaceTags, onToggleTag, onCreateTag, getTagById }: NoteTagBarProps) {
-  const [pickerOpen, setPickerOpen] = useState(false);
-  const [newTagName, setNewTagName] = useState("");
-  const pickerRef = useRef<HTMLDivElement>(null);
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  // Close picker when clicking outside
-  useEffect(() => {
-    if (!pickerOpen) return;
-    function handleClick(e: MouseEvent) {
-      if (pickerRef.current && !pickerRef.current.contains(e.target as globalThis.Node)) {
-        setPickerOpen(false);
-        setNewTagName("");
-      }
-    }
-    document.addEventListener("mousedown", handleClick);
-    return () => document.removeEventListener("mousedown", handleClick);
-  }, [pickerOpen]);
-
-  useEffect(() => {
-    if (pickerOpen) inputRef.current?.focus();
-  }, [pickerOpen]);
-
-  const assignedTags = note.tagIds.map((id) => getTagById(id)).filter(Boolean) as import("@/types").Tag[];
-  const unassignedTags = workspaceTags.filter((t) => !note.tagIds.includes(t.id));
-  const filteredUnassigned = newTagName
-    ? unassignedTags.filter((t) => t.name.toLowerCase().includes(newTagName.toLowerCase()))
-    : unassignedTags;
-
-  function handleCreateTag() {
-    const trimmed = newTagName.trim();
-    if (!trimmed) return;
-    onCreateTag(trimmed);
-    setNewTagName("");
-    setPickerOpen(false);
-  }
-
-  return (
-    <div className="flex items-center gap-1.5 mt-2.5 max-w-4xl mx-auto flex-wrap relative">
-      {/* Assigned tags */}
-      {assignedTags.map((tag) => (
-        <button
-          key={tag.id}
-          onClick={() => onToggleTag(tag.id)}
-          className="group flex items-center gap-0.5"
-          title={`Remove tag "${tag.name}"`}
-        >
-          <Badge color={tag.color}>{tag.name}</Badge>
-          <X size={9} className="opacity-0 group-hover:opacity-100 text-[var(--text-tertiary)] transition-opacity -ml-1" />
-        </button>
-      ))}
-
-      {/* Add tag button */}
-      <div className="relative" ref={pickerRef}>
-        <button
-          onClick={() => setPickerOpen((o) => !o)}
-          className="flex items-center gap-1 px-1.5 py-0.5 rounded text-[0.714rem] text-[var(--text-tertiary)] hover:text-[var(--text-secondary)] hover:bg-[var(--surface-2)] border border-dashed border-[var(--border)] transition-colors"
-        >
-          <Tag size={11} />
-          Add tag
-        </button>
-
-        {pickerOpen && (
-          <div className="absolute top-full left-0 mt-1 z-20 bg-[var(--surface)] border border-[var(--border)] rounded-lg shadow-lg p-2 w-48">
-            <input
-              ref={inputRef}
-              value={newTagName}
-              onChange={(e) => setNewTagName(e.target.value)}
-              onKeyDown={(e) => {
-                if (e.key === "Enter") {
-                  if (filteredUnassigned.length === 0 && newTagName.trim()) handleCreateTag();
-                  else if (filteredUnassigned.length > 0) { onToggleTag(filteredUnassigned[0].id); setPickerOpen(false); setNewTagName(""); }
-                }
-                if (e.key === "Escape") { setPickerOpen(false); setNewTagName(""); }
-              }}
-              placeholder="Search or create…"
-              className="w-full px-2 py-1 text-xs rounded bg-[var(--surface-2)] border border-[var(--border)] text-[var(--text-primary)] placeholder:text-[var(--text-tertiary)] focus:outline-none focus:border-[var(--accent)] mb-2"
-            />
-            <div className="max-h-36 overflow-y-auto space-y-0.5">
-              {filteredUnassigned.map((tag) => (
-                <button
-                  key={tag.id}
-                  onClick={() => { onToggleTag(tag.id); setPickerOpen(false); setNewTagName(""); }}
-                  className="flex items-center gap-2 w-full px-2 py-1 rounded text-xs hover:bg-[var(--surface-2)] transition-colors text-left"
-                >
-                  <span className="w-2 h-2 rounded-full flex-shrink-0" style={{ backgroundColor: tag.color }} />
-                  <span className="text-[var(--text-secondary)] truncate">{tag.name}</span>
-                </button>
-              ))}
-              {newTagName.trim() && (
-                <button
-                  onClick={handleCreateTag}
-                  className="flex items-center gap-2 w-full px-2 py-1 rounded text-xs hover:bg-[var(--surface-2)] transition-colors text-left text-[var(--accent)]"
-                >
-                  <Plus size={10} />
-                  Create &quot;{newTagName.trim()}&quot;
-                </button>
-              )}
-              {filteredUnassigned.length === 0 && !newTagName.trim() && (
-                <p className="text-[0.786rem] text-[var(--text-tertiary)] px-2 py-1">No tags yet</p>
-              )}
-            </div>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-/** Count words in markdown content (strips syntax first) */
-function countWords(md: string): number {
-  const text = md
-    .replace(/```[\s\S]*?```/g, "")
-    .replace(/`[^`]*`/g, "")
-    .replace(/^#{1,6}\s+/gm, "")
-    .replace(/[*_~]/g, "")
-    .replace(/\[([^\]]+)\]\([^)]+\)/g, "$1")
-    .replace(/!\[[^\]]*\]\([^)]+\)/g, "")
-    .replace(/https?:\/\/\S+/g, "")
-    .trim();
-  if (!text) return 0;
-  return text.split(/\s+/).filter(Boolean).length;
-}
-
-/** Strip markdown syntax to get plain text for search indexing */
-function stripMarkdown(md: string): string {
-  return md
-    .replace(/^#{1,6}\s+/gm, "")
-    .replace(/\*\*(.+?)\*\*/g, "$1")
-    .replace(/\*(.+?)\*/g, "$1")
-    .replace(/`{1,3}[^`]*`{1,3}/g, "")
-    .replace(/\[([^\]]+)\]\([^)]+\)/g, "$1")
-    .replace(/^[-*+]\s+/gm, "")
-    .replace(/^\d+\.\s+/gm, "")
-    .replace(/^>\s+/gm, "")
-    .replace(/^---+$/gm, "")
-    .replace(/\|/g, " ")
-    .replace(/\n{2,}/g, "\n")
-    .trim();
-}
+     </div>
+   );
+ }
